# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RoleEventAction = typeof RoleEventAction[keyof typeof RoleEventAction];


export const RoleEventAction = {
  assigned: 'assigned',
  removed: 'removed',
} as const;

export interface RoleEvent {
  id: number;
  guildId: string;
  executorId: string;
  executorTag: string;
  targetId: string;
  targetTag: string;
  roleId: string;
  roleName: string;
  action: RoleEventAction;
  assignedAt: string;
}

export interface RoleEventPage {
  data: RoleEvent[];
  total: number;
  page: number;
  limit: number;
}

export interface TopMod {
  executorId: string;
  executorTag: string;
  assigned: number;
  removed: number;
  total: number;
}

export interface TopRole {
  roleName: string;
  times: number;
}

export interface RoleStats {
  totalAssigned: number;
  totalRemoved: number;
  totalEvents: number;
  topMods: TopMod[];
  topRoles: TopRole[];
}

export interface Guild {
  id: string;
  name?: string;
}

export interface GuildList {
  guilds: Guild[];
}

export type ListRoleEventsParams = {
page?: number;
limit?: number;
guildId?: string;
action?: ListRoleEventsAction;
search?: string;
};

export type ListRoleEventsAction = typeof ListRoleEventsAction[keyof typeof ListRoleEventsAction];


export const ListRoleEventsAction = {
  assigned: 'assigned',
  removed: 'removed',
} as const;

export type GetRoleStatsParams = {
guildId?: string;
};