# lib/api-zod/src/generated/types/guildList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Guild } from './guild';

export interface GuildList {
  guilds: Guild[];
}