# lib/api-zod/src/generated/types/roleEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoleEventAction } from './roleEventAction';

export interface RoleEvent {
  id: number;
  guildId: string;
  executorId: string;
  executorTag: string;
  targetId: string;
  targetTag: string;
  roleId: string;
  roleName: string;
  action: RoleEventAction;
  assignedAt: Date;
}