# lib/api-zod/src/generated/types/roleEventPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoleEvent } from './roleEvent';

export interface RoleEventPage {
  data: RoleEvent[];
  total: number;
  page: number;
  limit: number;
}