# lib/api-zod/src/generated/types/roleStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TopMod } from './topMod';
import type { TopRole } from './topRole';

export interface RoleStats {
  totalAssigned: number;
  totalRemoved: number;
  totalEvents: number;
  topMods: TopMod[];
  topRoles: TopRole[];
}